+++ src/main/java/com/wedit/backend/common/config/security/SecurityConfig.java
@@ -1,7 +1,10 @@
 package com.wedit.backend.common.config.security;
 
 import com.wedit.backend.api.member.jwt.filter.FilterExceptionHandler;
-import com.wedit.backend.common.config.jwt.JwtConfig;
+import com.wedit.backend.api.member.jwt.filter.JwtAuthenticationProcessingFilter;
+import com.wedit.backend.api.member.jwt.repository.RefreshTokenRepository;
+import com.wedit.backend.api.member.jwt.service.JwtService;
+import com.wedit.backend.api.member.repository.MemberRepository;
 import com.wedit.backend.common.oauth2.OAuth2AuthenticationFailureHandler;
 import com.wedit.backend.common.oauth2.OAuth2AuthenticationSuccessHandler;
 import com.wedit.backend.common.oauth2.OAuth2UserService;
@@ -27,11 +30,14 @@ import java.util.Arrays;
 @RequiredArgsConstructor
 public class SecurityConfig {
 
-    private final JwtConfig jwtConfig;
     private final FilterExceptionHandler filterExceptionHandler;
     private final OAuth2AuthenticationSuccessHandler oAuth2AuthenticationSuccessHandler;
     private final OAuth2AuthenticationFailureHandler oAuth2AuthenticationFailureHandler;
     private final OAuth2UserService oAuth2UserService;
+
+    private final JwtService jwtService;
+    private final MemberRepository memberRepository;
+    private final RefreshTokenRepository refreshTokenRepository;
 
     public static final String[] PERMIT_URL_ARRAY = {
             "/api/swagger-resources/**",
@@ -53,6 +59,12 @@ public class SecurityConfig {
             "/api/v1/member/**",
             "/api/v1/vendor/**"
     };
+
+
+    @Bean
+    public JwtAuthenticationProcessingFilter jwtAuthenticationProcessingFilter() {
+        return new JwtAuthenticationProcessingFilter(jwtService, memberRepository, refreshTokenRepository);
+    }
 
     @Bean
     public PasswordEncoder passwordEncoder() {
@@ -130,7 +142,7 @@ public class SecurityConfig {
                         .accessDeniedHandler(filterExceptionHandler)        // 인가 실패 예외 핸들링
                 );
 
-        http.addFilterBefore(jwtConfig.jwtAuthenticationProcessingFilter(), UsernamePasswordAuthenticationFilter.class);
+        http.addFilterBefore(jwtAuthenticationProcessingFilter(), UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
     }
